requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
MouseInfo==0.1.3
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
setuptools==74.1.2
six==1.16.0
